Separate functional tests from the rest
Travis only provides go 1.4 in the Python test environment. To use
1.7, you have to use a Go test environment. Follow kinto-signer in
extracting this sort of test into a separate file and running it in a
separate environment.
