+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py33, py34, py35, py36, flake8
+envlist = py33, py34, py35, py36, flake8, functional
 
 [travis]
 python =
@@ -22,3 +22,11 @@ deps =
 commands =
     pip install -U pip
     py.test --basetemp={envtmpdir}
+
+[testenv:functional]
+basepython = python3
+deps =
+    -r{toxinidir}/requirements_dev.txt
+    -r{toxinidir}/aws_lambda/requirements.txt
+commands =
+    py.test tests/functional.py
